Added HTTP Post integration tests
