+++ SmtpToRest.IntegrationTests/Services/Smtp/SmtpServerBackgroundServiceTests_Basic.cs
@@ -78,4 +78,79 @@ public partial class SmtpServerBackgroundServiceTests
 		Assert.NotNull(result);
 		result.IsSuccess.Should().BeTrue();
 	}
+
+	[Fact]
+	[Trait(CategoryKey, CategoryBasic)]
+	public void ProcessMessages_ShouldIncludeApiToken_WhenSuppliedInConfiguration()
+	{
+		// Arrange
+		Configuration.HttpMethod = Rest.HttpMethod.Get.ToString();
+		Configuration.ApiToken = "someToken";
+		ConfigurationMapping mapping = new();
+		Mock<IMimeMessage> message = Arrange("sender@somewhere.com", mapping);
+		HttpMessageHandler
+			.Expect(HttpMethod.Get, Configuration.Endpoint)
+			.With(r => r.Headers?.Authorization is { Scheme: "Bearer", Parameter: "someToken" })
+			.Respond(HttpStatusCode.OK);
+
+		// Act
+		ProcessResult? result = SendMessage(message.Object);
+
+		// Assert
+		HttpMessageHandler.VerifyNoOutstandingExpectation();
+		Assert.NotNull(result);
+		result.IsSuccess.Should().BeTrue();
+	}
+
+	[Fact]
+	[Trait(CategoryKey, CategoryBasic)]
+	public void ProcessMessages_ShouldRespectOverriddenValues_WhenOverriddenInMapping()
+	{
+		// Arrange
+		Configuration.HttpMethod = Rest.HttpMethod.Get.ToString();
+		Configuration.ApiToken = "someToken";
+		ConfigurationMapping mapping = new()
+		{
+			CustomHttpMethod = Rest.HttpMethod.Post.ToString(),
+			CustomEndpoint = "http://overriddenendpoint",
+			CustomApiToken = "someOverriddenToken"
+		};
+		Mock<IMimeMessage> message = Arrange("sender@somewhere.com", mapping);
+		HttpMessageHandler
+			.Expect(HttpMethod.Post, "http://overriddenendpoint")
+			.With(r => r.Headers?.Authorization is { Scheme: "Bearer", Parameter: "someOverriddenToken" })
+			.Respond(HttpStatusCode.OK);
+
+		// Act
+		ProcessResult? result = SendMessage(message.Object);
+
+		// Assert
+		HttpMessageHandler.VerifyNoOutstandingExpectation();
+		Assert.NotNull(result);
+		result.IsSuccess.Should().BeTrue();
+	}
+
+	[Fact]
+	[Trait(CategoryKey, CategoryBasic)]
+	public void ProcessMessages_ShouldAppendService_WhenSuppliedInMapping()
+	{
+		// Arrange
+		Configuration.HttpMethod = Rest.HttpMethod.Get.ToString();
+		ConfigurationMapping mapping = new()
+		{
+			Service = "someService"
+		};
+		Mock<IMimeMessage> message = Arrange("sender@somewhere.com", mapping);
+		HttpMessageHandler
+			.Expect(HttpMethod.Get, $"{Configuration.Endpoint}/someService")
+			.Respond(HttpStatusCode.OK);
+
+		// Act
+		ProcessResult? result = SendMessage(message.Object);
+
+		// Assert
+		HttpMessageHandler.VerifyNoOutstandingExpectation();
+		Assert.NotNull(result);
+		result.IsSuccess.Should().BeTrue();
+	}
 }
